+++ sort.h
@@ -4,13 +4,7 @@
 #include "structs.h"
 #include <stdio.h>
 
-//void Sort(void* text, size_t size_el, int l, int r, int (*compare) (void* s1, void* s2));
-
-//void Copy(void* to, void* from, size_t size_el);
-
-//void Qsort(void *a, size_t size_el, int l, int r, int (*comp_func)(void *a, void *b));
-
-void Qsort(void* a, size_t size_el, int l, int r, int (*comp_func)(void *a, void *b));
+void Qsort(void* a, size_t size_el, int l, int r, long long (*comp_func)(void *a, void *b));
 
 void Swap(void* to, void* from, size_t size_el);
